add pydantic
