+++ pyproject.toml
@@ -3,15 +3,20 @@ name = "api-workshop"
 version = "0.1.0"
 description = ""
 authors = [
-    {name = "Gabriel",email = "gbuzak@gmail.com"}
+    {name = "Gabriel", email = "gbuzak@gmail.com"}
 ]
 readme = "README.md"
-requires-python = ">=3.13"
+requires-python = ">=3.13,<4.0"
 dependencies = [
-    "pytest (>=8.3.4,<9.0.0)"
+    "pytest (>=8.3.4,<9.0.0)",
+    "taskipy (>=1.14.1,<2.0.0)"
 ]
-
 
 [build-system]
 requires = ["poetry-core>=2.0.0,<3.0.0"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.taskipy.tasks]
+run = "uvicorn app.main:app --reload"
+test = "pytest -v"
+kill = "kill -9 $(lsof -t -i:8000)"
